default to calculate deps-graph from the saved flattened-edges unless --rebuild-deps-graph was used (#8301)
This is a continuation of https://github.com/teambit/bit/pull/8281.
In that PR this calculation was done only when an env variable was used.
This PR takes it to the next step, which enables it by default. In case
an error was thrown due to discrepancies between the methods of
calculating the graph, it suggests adding `--rebuild-deps-graph` flag to
not get blocked.