fix(diverge-data), find the common-snap closer to both source and target (#8317)
Currently, when calculating the common-snap, we find all common snaps
and then traversing BFS on the source sub-graph and return the first
one.
This is good for most cases. However, in some cases, traversing BFS for
both the source and target, yield a different snap.
This PR changes this algorithm to find out which one is a successor of
the other and returns the one closer to the head. (the parent).