fix(lane-merge), avoid adding non-lane components into the current lane when merging from main (#8337)
Due to a recent change https://github.com/teambit/bit/pull/8279, the
default when running `bit lane merge main` is to include non-lane
components.
The issue is that these non-lane components were also added to the lane
object. So for example, if a workspace has 10 components and only 2 were
part of the lane, after the merge, all 10 components became part of the
lane.

This PR fixes it by updating non-lane components in the
.bitmap/file-system but not adding them to the lane object.
If the flag `--exclude-non-lane-comps` is used, then it doesn't touch
these non-lane components.