feat(lanes), introduce lane history (#8370)
A lane object can change due to snap/merge/import. Currently there is no
way to track changes to a lane object. Once the lane is changed, it's
impossible to know how it looked like previously.

This PR implements the infrastructure of the Lane History feature.
* introduce a new `BitObject` object: `LaneHistory`.
* once a change is done to lane object (for example, when snapping
components when on a lane), it updates this LaneHistory object. It
creates a uuid for a unique key to this change. The values it stores are
the author, date, the reason for the change and the components map
(id:head).
* when a lane is exported, it also exports the LaneHistory object.
* on the remote, when a LaneHistory object is received, it is merged
with the existing LaneHistory object. (it's a simple merge - only add
new keys).
* introduce a new command `bit lane history <lane-name> [id]`. 

Currently, it can't export this LaneHistory object because the remotes
need to be updated with this new object-type. Once the remotes are
updated, it can be exported by using a bit-feature flag
`support-lane-history`. Also the new command is hidden behind this
feature flag. This flag won't be needed once we verified that all remote
scopes are up to date.