fix(snap), fix ComponentNotFound error thrown during isPartOfMainHistory check (#8412)
This is a regression caused by a recent optimization of not importing
flattened-deps during snap.
One of the validations occurring during snap/tag on lane is that all
direct dependencies are either part of the same lane or main. To check
this we need the `ModelComponent` object, which was missing due to the
optimization above.
This PR fixes it by adding the missing one to the "missingDeps" array
which later gets imported.