fix(reset), remove from lane if was reset to main and not exist on remote-lane (#8454)
Here is a scenario where this was an issue: `scope-a/comp` was imported
to the workspace, a new lane was created on another scope: `scope-b/dev`
and this component was changed and snapped so it became part of the
lane.
Then, `bit reset` was running which changed the head on the lane to be
the same as main. Now `bit export` throws an error saying `scope-a/comp`
doesn't exist. This is because the lane-object included this component
but this component wasn't sent to the remote as it wasn't staged.

This PR fixes it by removing the component from the lane during the
reset command. Because the component doesn't exist on the remote-lane
and its head is the same as main, no reason to keep it on the lane.