fix(deps-set), add snap prefix (0.0.0-) to the version in .bitmap if needed (#8472)
`bit deps set` saves the data into the `policy` of the dependencies
resolver config, which accepts package name and a valid Semver. Users
are not aware of the prefix added to the package version of snaps, so
running `bit deps set pkg-name@snap-hash` result in saving the version
as the snap-hash in the objects, which later fails the installation.
This PR adds the `0.0.0-` prefix to the snap hash to make it a valid
Semver.