fix(update-dependencies), add flattened-dependencies when --simulation is used (#8494)
Otherwise, in case of updating envs/aspects, the version-validator
throws an error saying "the extension ${extensionId.toString()} is
missing from the flattenedDependencies".

We didn't add flattened-deps in the past because it was very slow, it
used to import all flattened dependencies. This is not the case anymore.
A recent change (a month ago) calculates the flattened using the
"flattened-edges" saved in the Version object.