feat(import), prompt available paths from imported components to dependents (#8505)
The flag `--dependents` of `bit import` brings all the components linked
from the workspace components to the specified component (target). In
many cases it's not needed and the user is interested in a specific path
from the workspace to the component.
This PR makes the following changes:
* if it found 10 paths or less, it shows a multi-select prompt with
available paths.
* if it found more than 10 paths, it prompts the first level only
(workspace-components) and once selected, it shows the next level only
until it reaches the specified component (target).
* once the paths are selected by the user, it shows a prompt with the
components that should be imported for the selected paths. The user can
confirm or abort.
* the `--dependents-dry-run` is deprecated. Now it's the default. It
always asks for confirmation before bringing in the dependents.
* the `--dependents-through` flag is renamed to `--dependents-via`.
* a new flag `--silent` is introduced to avoid all prompts of
`--dependents` and `--dependents-via`