feat(merge-from-scope), support merging main into lane (#8511)
Until now, merge-from-scope was able to do a simple merge, e.g. lane to
main when the lane is up do date, so all it did was changing the head
and its parents.
Merging main into a lane is more complex because normally the main and
the lane are diverged. The classic use-case is creating a CR, planning
to merge to main but then main progressed with more snaps so the lane
became out of date. In this use case, main and lane are diverged, so it
needs a "snap-merge" to point to both main and lane, and it needs
files-merge to check whether there are conflicts, in which case we stop
the merge and notify the user.