feat, implement update-dependents on lane to include dependents in the lane but skip them in the workspace (#8534)
These components will be added to the lane into a dedicated property
"updateDependents", which is safer than adding them to `components`
property with a flag. This way we know they won't get into places we
didn't intended them to go, such as the user workspace or importing
their objects.
The main purpose of this is to see in the UI how the dependents work
with these components of the lane, and if all good, once merged to main,
these are merged as well.