fix, handle the case when importing a component from updateDependents into the workspace (#8538)
When checked out to a lane and running `bit import <id>` of a component
included in the `updateDependents` prop of the lane, don't bring it from
there. Fetch it from main instead. Only once it's snapped, add it to the
lane. It's ok that in this point, the component exists in both:
`components` array and `updateDependents` array of the lane. Later, once
this lane object is merged into the remote, it'll be removed from the
`updateDependents` array and it'll be only in `components`. This removal
needs to be done on the remote, because only the remote makes changes in
this array unless it gets "overrideUpdateDependents" flag, which enabled
by snap-from-scope only.