fix(merge-from-scope), do not throw an error about missing artifacts when merging main to a lane (#8557)
Previously, `collectVersionsObjects` had a bug,
`ignoreMissingLocalArtifacts` didn't do anything, even when it was
`false`, it was still ignoring local artifacts because
`locallyChangedHashes.includes(versionObject.hash.toString())` always
returned `false`. The `versionObject.hash.toString()` printed the hash
method instead of the hash content. It's fixed to be
`versionObject.hash().toString()` to make it work.
Because of the that bug, https://github.com/teambit/bit/pull/7119 was
adding new flag `ignoreMissingExternalArtifacts` unnecessarily. It
assumed that an error wasn't thrown because it was external which wasn't
correct. In this PR this flag has removed.
In addition to this, when merging from scope and merging main to a lane,
we can skip the import of the artifacts because once the user asks for
the artifacts we retrieve it from main if not found in the lane.

One last thing, because until now `ignoreMissingLocalArtifacts`
practically did nothing, I'm afraid that after this fix it'll throw
errors unnecessarily, so I made the default to not throw when missing.
We didn't get any complain about missing artifacts in the remote scopes,
so it should be fine.
This flag was renamed to `throwForMissingLocalArtifacts` for clarity.
The only place we throw is when merging-from-scope and the source is a
lane, not main.