fix(diff), do not show diff when it differs only in the EOF char (#8610)
This fixes a bug where an svg file downloaded from a Win machine was
repeatedly shown as modified on Mac. Even after tag/snap/checkout-reset.
The reason is that when saving files into the objects, their EOF is
converted to Linux (`\n`). The files on the FS on the other hand, stay
the same. If they've `\r\n`, it'll stay this way. Then when sending both
files to git to check the diff, it shows all lines as diff.
This PR converts the `SourceFile` into a `Source` object which takes
care of the Linux conversion and produces a hash, then if the hash is
the same, it doesn't send them to git to check the diff.