fix(lane-merge), avoid removing dependencies that were set explicitly before (#8665)
In case a lane-merge introduces some dependencies changes, it used to
disregard the "bit deps set" that were saved previously. It was handling
only the "auto detect", and because the policy is not empty, the next
strategy - "model specific" was ignored. As a result, any previous "deps
set" was ignored.
This PR fixes it to add these dependencies to the merge-config.