prepare converting flattened-edges into strings (#8690)
currently, the flattened-edges are saved as an array of objects.
with this change, it'll be saved as a tuple array `[source, target, type]`.
I tested this in a component with a huge dependency graph of 604 edges.
The file size has shrunk from 143K to 6.1K !!

The only issue is that this object file is parsed elsewhere with old bit
version which expect it to be an object.
That's why in this PR we don't do the switch yet. We support both
formats when parsing, but we save it with the old format.
In a few months we could make the switch.

One more change in this PR is the ability to use `bit cat-object -p `
for arbitrary source files and if the file is a JSON file, it'll be
formatted nicely.