fix(compile), avoid showing the same component-id multiple times (#8701)
It's easy to reproduce it by running `bit create node-env my-env && bit
compile`.
Until now it used a pub-sub mechanism for the compilation events. In
this case, because `bit install` wasn't running yet, the aspect failed
to load and fallback to re-compile. Practically, compiling the component
twice, sending two events and then reporting these two events to the
user.
In this PR, the event of completing the compile has been removed. It
gets the results directly instead of using events.