make watch output more compact, refactor formatCompileResults to use toString() (#8706)
This `formatCompileResults` was implemented incorrectly. The watcher was
building the output of the compiler based on unknown data.
In this PR, the compiler formats its own data and returns the formatted
data to the `toString()` method in `OnComponentEventResult` type as
intended originally. This way, each aspect can have its own formatting
and as long as it implements `toString()`, it can be shown in the
output.