improvement(version-history), enable generating graph as SVG without installing graphviz (#8714)
Until now to generate a graph (either for Version History or for
dependencies), the GraphViz library had to be installed in the OS, which
made it cumbersome especially for Windows.
This PR uses [viz-js](https://github.com/mdaines/viz-js) package which
is basically an implementation of Graphviz in javascript. It doesn't
support png/gif formats, but it does support "svg" which is superior for
such graphs that can be huge. (with png trying to zoom in result in a
blurry image).
In this PR, the `bit version-history graph` command defaults to this new
alternative and a new flag `--png` falls back to GraphViz.