fix(merge-from-scope), merge components that were deleted with --update-main (#8763)
Currently the code that merges from scope is different than the one on
the workspace (except when merging main to a lane).
As a result, the change done in the last few months of allowing
`--update-main` flag during bit-delete was supported only in the
workspace. On the scope, it was just ignoring the component, assuming it
was deleted and should not be merged.
This PR reuses the merge on the workspace instead of having different
code on the scope, which as a result, merges those deleted components
with "update-main" prop.