allow exiting (Ctrl+C) during typescript-compiler step in build pipeline (#8775)
When registering to process.on with exit signals, Typescript doesn’t
exit the process until it executes the code that registered to this
event.
Because typescript programatic API is sync and JS is single threaded,
it’s blocked until TS completes the build and only then it can execute
the code registered to the exit event.
In the past, Ora (the spinner) pkg caused it and it was solved by
instantiating it with `hideCursor: false`. Tuned out the reason it fixed
it is because this option when it’s true, it calls `cliCursor` package
to hide the cursor. This package, is calling `restore-cursor` package
which calls `signal-exit` package to register to exit events to make
sure the cursor is restored when exiting the node process.

The reason we had the same issue now is that we had two places we
registered to the exit signal.
1. when the logger starts, it calls `spinnies = new Spinnies();` which
comes from `dreidels` pkg which calls to cliCursor same as Ora, but this
pkg doesn’t have an option to opt-out. It's fixed to insatiate it
lazily. (currently only `bit start` uses it).
2. in the `build-pipe.ts` before running the pipeline, it calls
`addSignalListener` which explicitly adding the exit listeners.
Absurdly, it was added to overcome this Ctrl+C issue, but it actually
helped to make it non-working. This is now removed.

This fix should help for any other blocking (sync) build-tasks. Not only
for typescript.