feat(watch), introduce --import to fetch objects on .bitmap change (#8795)
Helpful when using git.
It resolves the following issue - a user is running `git pull` which
updates the components and the .bitmap file. Because the objects locally
are not updated, the .bitmap has new versions that don't exist in the
local scope. As soon as the watcher gets an event about a file change,
it loads the component which throws `ComponentsPendingImport` error.

To resolve this, the new `--import` flag imports the new objects as soon
as the .bitmap file changes. For performance reasons, we import only
when: 1) the .bitmap file has version changes and 2) this new version is
not already in the scope.