fix(lane-merge), remove a package from workspace.jsonc policy if newly introduced during merge (#8817)
The scenario is as follows:
Lane-A has comp-X in the workspace.jsonc policy, it is not part of the
lane.
Lane-B has this comp-X as part of the lane. 
Lane-B is now merged into Lane-A.
This PR removes the entry of comp-x package from workspace.jsonc, so
then this component doesn't exist in both: .bitmap and workspace.jsonc.