shorten the capsule dir (#8827)
The capsule base-dir is normally a `sha1` of the workspace/scope dir,
which is 40 chars long. Problem is that Windows is not happy with long
paths.
This PR uses a shorter hash. The number 9 is pretty random, it's what we
use for short-hash of snaps.
We're aware of an extremely low risk of collision. Keep in mind that in
most cases you won't have more than 10 capsules in the machine, so it's
fine.