fix(checkout-reset), reset component.json if exists locally (#8826)
currently, if there are local changes in the component.json (generated
by `bit config eject-conf`), then `bit checkout reset` doesn't reset
them. It leaves the file as is.
This PR refactors the way how this component.json file is saved to reuse
the infrastructure we had for package.json (with indent and EOL taken
into account). With this, it was easier to change the component-written
command to make sure this file is written if it was there before.