feat(lanes), introduce a new sub-command "merge-move" (#8834)
This command is useful when you got a messy merge state that from one
hand you don't want to loose the changes, but on the other hand, you
want to keep your lane without those changes.
This command does the following:
1. create a new lane with the current merge state. including all the
filesystem changes. (in practice, it leaves the fs intact).
2. reset the current lane to the state before the merge, so then once
done with the new lane, you can switch to the current lane and it'll be
clean.