improvement(import), allow importing main components with wildcard when on a lane (#8836)
We used to throw an error suggesting the user to import the components
individually. The reason was that when you're on a lane, you probably
want to import components from the lane. not main. That's why when the
wildcard matches both, lane and main, only the lane are imported.
However, if only components from main are matched, it's not very clear
whether you really want to import from lane, or your wildcard syntax
didn't bring the results you thought it would.
With this PR we assume that in this case you do want to import from main
so we let you do that.