feat(delete), introduce --range flag to delete tags of a specific range (#8844)
Similar to `bit deprecate --range`, this new flag allows deleting
specific tags. For example `bit delete my-comp --range 0.0.1`. It marks
0.0.1 only as deleted.
This is supported in `bit log`. It's also shown in the `bit import` and `bit show` output. 
For example, when running `bit show my-comp@0.0.2` the "removed" is false. 
However, `bit show my-comp@0.0.1` shows it as true. It also shows the range.