feat, support using .gitignore file in the component root-dir (#8847)
Until now, `.gitignore` files were always ignored by default and never
tracked by Bit.

## Proposed Changes

- track `.gitignore` files. This way, when multiple users import the
same component and using different git repos, they could use the same
.gitignore file.
- in case `.gitignore` is found in the component root-dir, Bit reads it
and uses it to untrack component files.
- in case `.bitignore` is found in the component root-dir in addition to
`.gitignore`, the `.bitignore` takes precedence and is used to untrack
component files. (still, `.gitignore` itself is tracked). This way, it's
possible to have different patterns/rules for bit and git.
- until now, bit was searching for the `.gitignore` of the
workspace-root, now only in the workspace-root, but also upwards until
it finds it. This is unintuitive and can cause confusion. It was changed
to use only the one in the workspace root.