add "loadAspects" to the Command class to allow skipping this step (#8849)
This is about the loadAspects triggered from the Workspace during the
onStart hook, which loads the aspects set on the workspace.jsonc file.
It takes time and if it's possible to avoid it, it improves the
performance. (on a new react workspace, it saves as much as 500ms).
The reason this is needed is that users can add aspects which register
to various slots and if these aspects are not loaded, they won't be
registered. Also, keep in mind, that these aspects are not dependencies
of the components, so the fact that workspace components are loaded,
doesn't mean that these aspects will be loaded.

Until now, it was skipped for bit login/logout. It was hard-coded. 
With this PR, the machanism is baked into `Component` class. Also, it is
skipped for all legacy commands. These commands don't use any slots, so
it's fine to not load the aspects for them.