introduced a new component-issue "DeprecatedDependencies" (#8854)
Useful to get noticed when workspace components are using deprecated
dependencies.
This component-issue is set to not be a tag blocker.

Also, made an improvement to the "issues found" text in bit-status to be
shown only when the issue is a tag-blocker. In case the issue is not a
blocker, it shows "warnings found".