change Version hash to be auto-generated and not based on content (#8887)
In the legacy it was completely based on hashing part of the content.
Since Harmony, only tag (not snap) used this calculation and only for
the first time. Once generated, the hash was saved as part of the object
and was used purely by the hash prop and not by the calculation.

This PR completes the transition by aligning tag and snap and always
auto-generating the hash by a random UUID.

(the reason this was done now is because I have noticed a very rare
occurrence when an e2e-test was creating two components with the exact
same files and their tag hashes were identical. It's rare because part
of the hash calculation based on the "log.date" which changes every
second).