fix(snap-from-scope), support dependents recursively with --update-dependents (#8937)
Easier to explain with example: comp1 -> comp2 -> comp3. Snapping comp2
with `--update-dependents` saves comp2 inside the `updateDependents`
prop of the lane with the comp3 dependency having the same version as in
the lane.
Snapping comp3 now have the same effect. It saves comp2 with the version
is has in the `updateDependents` prop.