fix(snap), save app data during snap/tag (#9037)
Until now, importing an app to a new workspace and snapping it, the
`Version` object is missing the data of the application. (specifically
`appName` prop).
This is because of the way the data is calculated, which relies on the
appSlot to have all apps registered. When the app is not in
workspace.jsonc it never registered.
This PR fixes it by loading the loading the app components within the
components to snap as aspects so then they're registered to the app slot
and recognized as apps.