feat, enable running bit commands using the background http bit-server (#9076)
To enable this, add an env variable `BIT_CLI_SERVER=true`.
This is how it works:
At bootstrap, bit checks whether the env variable above exists and if so
it doesn't invoke "yargs" to parse the arguments in the command line.
Instead, it checks whether there is bit-server running in the background
and if so it sends an http request with the cli arguments. Bit server in
the background gets the request, parse the arguments, invoke the command
and returns the response.