change the way of interacting with cli-server, write to the client tty directly on mac/linux (#9104)
Linux OS family treat the tty (terminal) as any other file, so it has an
actual file path we can write into.
When a client runs a bit command (using the env variable of enabling the
cli-server) it sends its tty path to the server. The server (in the
cli-raw route) monkey patches (no choice here) the
`process.stdout.write` and adds additional write - to the tty path. This
way, the client gets all the data stream in between the request and the
response.