feat(merge-lane)!: change no-snap to avoid fast-forward (#9159)
Until now, in case lane-b is merged into lane-a and a component (e.g.
comp-a) is ahead in lane-b, we simply updated lane-a with the new snap
of lane-b. (this behavior is similar to the "fast-forward" in Git).
This was true regardless of `--no-snap` flag. The `--no-snap` flag was
only for cases when comp-a is diverged in history and no conflicts
found, which was auto-snapping the component and saving two parents, the
head of lane-a and the head of lane-b.

In come cases, the end result of such a merge is not desirable. Some
changes might be done to this comp-a in lane-b which I would like to
investigate and test before I decide to get them in lane-a. Although
comp-a wasn't changed/snapped in my lane-a, still, I prefer to get the
changes of comp-a as modified, so then `bit diff` will show the new
changes I'm getting and then I can decide later to snap these changes.
This PR makes this possible by using the `--no-snap` flag. In practice,
it considers this state of target-ahead (lane-b) as diverged, which, as
a result - doesn't update lane-a, only the component-files and the
unmerged files with the heads data.

A new flag `--no-auto-snap` was introduced to keep the previous behavior
of `--no-snap` in case users need them.