fix(export), fix output to show the exported lane in case no new snaps added (#9175)
In some cases, the export doesn't bring any new component or snap/tag to
the remote, but it does update the lane object. It happens for example
in the following two cases:
1. when a lane has forked from another lane and exported immediately to
the same scope, without adding new snaps.
2. when merging another lane where all snaps already in the remote.

Until now, it was showing "nothing to export" for both cases above. The
first case actually didn't do any action and nothing was exported, the
second case the message is misleading because the lane object itself was
exported.

This PR fix both cases. The first case is now exporting the new lane to
the remote and for both cases it shows the message "exported the lane
xxx/yyy" in the output.