fix, always try to load objects from repo.objects before going to cache or fs (#9198)
See a previous PR for more context -
https://github.com/teambit/bit/pull/6404/files.
It's about tag/snap where the component version is changed to a future
version and later on during the process the workspace/scope asking for
this component by the ID.
Turned out that in some cases this is still an issue. It fails in
different places than the PR above.
In this PR, instead of passing the `preferInMemoryObjects` argument in
more places, and then potentially failing in different places in the
future, we simply go to the `objects` array in all cases and only if not
found, we try other sources.