fix(snap-from-scope), set a lane as new in scope.json to avoid fetching it from a remote (#9202)
The bug happened when snapping from scope a component with dependencies
and creating a new lane. The `getLaneForFetcher` method was calling
`scope.getCurrentLaneObject()`, which re-load the lane object from the
filesystem and as a result, lost the `isNew: true` prop.
This PR, mimics the behavior on the workspace, which is saving the info
about the lane being new in scope.json. This way, loading the lane
guarantees to have `isNew: true`.