fix, delete index.json on the remote if it is outdated (#9242)
This PR reverts https://github.com/teambit/bit/pull/1624.
In case the remote encounters an outdated index.json it wasn't throw an
error about it. The problem is that some functions calling it expect to
get objects. (e.g. `loadLane`). This PR, removes the check whether it's
a remote, and whenever an index.json file is outdated, it deletes it and
throw an error telling the user to re-run the command.