implement a new method of interacting with bit-server: node-pty (#9247)
Until now we supported two methods to communicate with cli-server: SSE
and TTY. See
[server-commander.ts](https://github.com/teambit/bit/pull/9247/files#diff-985a793a46f5dc18729617c89fafd82315c145a0f767760934b4f81912f3cc48)
for more details.

With this approach, we spawn a pty (pseudo terminal) process to
communicate between the client and the server.
The client connects to the server using a (TCP) socket. The server
writes to the socket and the client reads from it.
The client also writes to the socket and the server reads from it. See
server-forever.ts to understand better.
In order to pass terminal sequences, such as arrow keys or Ctrl+C, the
stdin of the client is set to raw mode.
In theory, this approach could work by spawning a normal process, not
pty, however, then, the stdin/stdout are non-tty, and as a result,
loaders such as Ora and chalk won't work.

With this new approach, we also support terminating and reloading the
server. A new command is added `bit server-forever`, which spawns the
pty-process. If the client hits Ctrl+C, this server-forever process will
kill the pty-process and re-load it.