feat, introduce "--stream" for --json commands to print loader as json strings (#9276)
This can be used by any command by simply adding `--stream` to the
`options` prop.
It's useful for processes that run bit commands with `--json` and need
to show some feedback to the user during the command run.
Normally, `--json` disables the loader to keep the output a valid JSON.
When adding this `--stream`, the loader is printed in JSON format with a
new line. The consuming process knows to consume one line at a time. The
final output includes `end: true` to indicate this is the last output of
the command.