feat, introduce a new private command "bit apply" (#9295)
This is helpful for code-generated to apply it easily with one command
to an existing component or to create a new component out of it.
It provides the ability to apply file/config changes to an existing
component or create a new component from the given files. Then, depends
on the passed flags, it can snaps/tags/exports. All within one command.