fix(reset), reintroduce a deleted component into .bitmap after reset (#9300)
Currently, when a component is deleted (`bit delete`), snapped and then
reset (`bit reset --all`), the reset ignores the deleted component. In
some cases, during the reset it throws ComponentNotFound error.
This PR includes the deleted components to be part of the reset, and
also makes sure to handle them differently. Their files are kept deleted
and it re-creates a .bitmap entry with the same props it had before the
snap.