improvement(watch), enable triggering onComponentChange on another component (#9301)
Introduced a new flag `--trigger <component-id>`, when any change is
happening, it runs the `workspace.onComponentChange` for the specified
component-id.
A real use case when this is needed is developing a vscode extension
component. Because a vscode-extension must be bundled, its dependencies
are always in the dist of the extension. As a result, when a dependency
in the workspace is changed, the extension is not aware of it and not
getting updated.
This flag is more like a workaround and not the best solution. Ideally,
the flag would be "trigger-dependents" which would traverse the graph
and trigger accordingly, but it's not worth it to implement as this
point as it's an edge use-case.