fix(snap), add the current checked out version as the parent, not the head (#9345)
Currently, in case the component is checked out to an older version, snap/tag command sets the parent of the new version as the head instead of the checked out version.
As a result, running `bit reset` changes `.bitmap` to the head version of the component rather than the version was used before the local snaps while the component files on the filesystem are according to the checked out version. This leads to a major confusion, because the changes between the head and the previously checked out version are gone once the user snaps again. 
This PR fixes it by using the checked out version as the parent. This leads to a component being diverged, because on the remote went to another direction since this checked out version. Once the user tries to export, it'll throw an error suggesting to either bit merge or bit reset and snap again.