decouple the teambit.harmony/bit dependency from core aspects (#9375)
Bit aspect (`teambit.harmony/bit`) depends on all core aspects. This is
needed because this aspect is the main entry of bit, it loads all those
core aspects and also we want to this aspect to get a new version every
time a core-aspect has changed, so then when running bit, all the core
aspect are always up to date.

However, some core aspects, e.g. `Generator` aspect were depended on Bit
aspect (practically forming a circular dependency) for mainly two tasks:
loadBit() and getting core-ids. Because of this circular, changing core
aspects triggers an auto-tag of most of the core aspects. Leading to
unnecessary tags and slow deployment.

This PR breaks this direction of dependency by injecting the APIs needed
to those aspects from the loadBit function. It's not pretty, but it's
much better than how it is now.