support maintaining multiple versions/branches on main (#9362)
In some scenarios, it's needed to maintain some legacy code, for
example, the new code is 2.x and the legacy is 1.x which still needs to
be maintained.
If a patch is needed to 1.x code, the component needs to be checked out
to 1.x and tagged with a specific version, for example: `bit tag --ver
1.0.1 --ignore-newest-version`.
Currently, when this patch is done, although the history of this patch
is correct thanks to a recent
[PR](https://github.com/teambit/bit/pull/9345), but the head now points
to this patch version. As a result, when importing the component, the
user gets 1.0.1 and not 2.x.
This PR introduces a new flag `--detach-head` to keep the head intact in
this case, so the head still points to 2.x, while the .bitmap is updated
to 1.0.1. This new patch 1.0.1 becomes a detached head.

It also supports lanes. It's possible to work on multiple components
when they're checked out to 1.x, create a new lane, snap some changes,
then merge the lane with `--detach-head` flag. This flag, instruct bit
to find the common snap before the divergent point, and merge the lane
to that point and not to the head.

In order for the garbage collector to not delete this detached-head
snap, and for the remote to not identify this component as a
merge-conflict, we save a new prop `detachedHeads` into `ModelComponent`
object.

For now, this flag `--detach-head` is under a feature-flag. It'll be
enabled by default once bit cloud is deployed with a new version that
supports it.